requirements.txt
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
numpy==2.0.1
opencv-python==4.10.0.84
pillow==10.4.0
pyzbar==0.1.9
requests==2.32.3
soupsieve==2.5
urllib3==2.2.2
